requirements.txt
PyQt6
Pillow
google-cloud-aiplatform
